Remove trade limits and boost supplier speed
Trade route resource limits are now set to INT16_MAX, effectively removing trade restrictions. Supplier movement speed has been increased by changing move ticks from 1 to 5. Land and sea trade unit limits for traders have been set to INT8_MAX instead of INT32_MAX.

+++ src/figuretype/trader.c
@@ -1019,7 +1019,7 @@ int figure_trade_land_trade_units(void)
         unit += add_unit;
     }
 
-    unit = INT32_MAX; // [rlaw]: unlimited land trade units for caravans and native traders
+    unit = INT8_MAX; // [rlaw]: unlimited land trade units for caravans and native traders
 
     return unit;
 }
@@ -1057,7 +1057,7 @@ int figure_trade_sea_trade_units(void)
         unit += add_unit;
     }
 
-    unit = INT32_MAX; // [rlaw]: unlimited land trade units for caravans and native traders
+    unit = INT8_MAX; // [rlaw]: unlimited land trade units for caravans and native traders
 
     return unit;
 }
